backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
python-multipart
pypdf
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
